refactor: Better redaction isolation
