+++ .mise-tasks/test/_default
@@ -3,9 +3,15 @@
 
 set -eux
 
-# Tests
+# Tests without features
 cargo nextest run --no-fail-fast
+
+# Tests with redaction feature
+cargo nextest run --no-fail-fast --features redaction
 
 # Tests in docs
 cargo test --doc
+
+# Tests in docs with redaction feature
+cargo test --doc --features redaction
