+++ .mise-tasks/test/review
@@ -1,4 +1,10 @@
 #!/usr/bin/env bash
 #MISE description="Run insta snapshot tests with nextest and review changes"
 
+set -eux
+
+# Run snapshot tests without features
 cargo insta test --test-runner nextest --review
+
+# Run snapshot tests with redaction feature
+cargo insta test --test-runner nextest --review --features redaction
