+++ Cargo.toml
@@ -13,7 +13,7 @@ homepage = "https://github.com/ilaborie/clawspec"
 serde = "1.0.219"
 serde_json = "1.0.140"
 serde_path_to_error = "0.1.17"
-serde_yaml = "0.9.34"
+serde_saphyr = { version = "0.0.7", package = "serde-saphyr" }
 serde_urlencoded = "0.7.1"
 
 utoipa = "5.3.1"
